trie: refactor lock class (#3109)
* trie: refactor lock

* trie: merge lock tests
